+++ sample/sample-with-di/app/src/main/java/com/dropbox/kaiken/sample/advancedsample/app/AdvancedKaikenSampleApplication.kt
@@ -1,9 +1,8 @@
 package com.dropbox.kaiken.sample.advancedsample.app
 
 import android.app.Application
-import com.dropbox.kaiken.skeleton.skeleton.core.SkeletonApplication
 import com.dropbox.kaiken.skeleton.skeleton.dagger.SdkSpec
-import com.dropbox.kaiken.skeleton.skeleton.usermanagement.UserStore
+import com.dropbox.kaiken.skeleton.usermanagement.UserStore
 import com.dropbox.kaiken.scoping.AppScope
 import com.dropbox.kaiken.scoping.SingleIn
 import com.dropbox.kaiken.scoping.SkeletonScope
