+++ skeleton/src/main/java/com/dropbox/kaiken/skeleton/usermanagement/ActiveUserDataSource.kt
@@ -1,4 +1,4 @@
-package com.dropbox.kaiken.skeleton.skeleton.usermanagement
+package com.dropbox.kaiken.skeleton.usermanagement
 
 import android.content.Context
 import android.content.SharedPreferences
