+++ skeleton/src/main/java/com/dropbox/kaiken/skeleton/usermanagement/UserDataSource.kt
@@ -1,4 +1,4 @@
-package com.dropbox.kaiken.skeleton.skeleton.usermanagement
+package com.dropbox.kaiken.skeleton.usermanagement
 
 import kotlinx.coroutines.flow.Flow
