--- skeleton/src/main/java/com/dropbox/kaiken/skeleton/usermanagement/auth/UserInput.kt
@@ -1,6 +1,0 @@
-package com.dropbox.kaiken.skeleton.usermanagement.auth
-
-/**
- * A holder class to feed the basic user information
- */
-data class UserInput(val userId: String, val username: String, val accessToken: String = "default")
